backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
celery==5.3.4
flower==2.0.1
redis==4.6.0
pydantic>=2.7.0
pydantic-settings>=2.0.3
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
openai>=1.66.2
openai-agents>=0.0.4
requests==2.31.0
pytest==7.3.1
httpx>=0.23.0,<0.24.0
supabase==1.0.3
gunicorn==21.2.0
python-jose==3.3.0
python-multipart==0.0.6
cryptography==41.0.3
PyJWT==2.6.0
